# project/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Coffee Shop Online Ordering API
  description: REST API for Coffee Shop Online Ordering System (Sprint 1 MVP)
  version: 1.0.0
paths:
...
  /api/v1/menu/categories:
    get:
      tags:
      - Menu & Catalog
      summary: List menu categories
      description: Retrieve all active menu categories ordered by display order.
      operationId: get_categories_api_v1_menu_categories_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CategoryResponse'
                type: array
                title: Response Get Categories Api V1 Menu Categories Get
    post:
      tags:
      - Menu & Catalog
      summary: Create menu category (admin only)
      description: Create a new menu category (requires admin privileges).
      operationId: create_category_api_v1_menu_categories_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/menu/categories/{category_id}:
    put:
      tags:
      - Menu & Catalog
      summary: Update menu category (admin only)
      description: Update an existing category (requires admin privileges).
      operationId: update_category_api_v1_menu_categories__category_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Menu & Catalog
      summary: Delete menu category (admin only)
      description: Delete a category (requires admin privileges).
      operationId: delete_category_api_v1_menu_categories__category_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menu/items:
    get:
      tags:
      - Menu & Catalog
      summary: List menu items
      description: Retrieve menu items with optional category filtering, availability,
        and search.
      operationId: get_menu_items_api_v1_menu_items_get
      parameters:
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Filter by category ID
          title: Category Id
        description: Filter by category ID
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search by name or description
          title: Search
        description: Search by name or description
      - name: available_only
        in: query
        required: false
        schema:
          type: boolean
          description: Filter only available items for public menu
          default: true
          title: Available Only
        description: Filter only available items for public menu
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItemResponse'
                title: Response Get Menu Items Api V1 Menu Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Menu & Catalog
      summary: Create menu item (admin only)
      description: Create a new menu item (requires admin privileges).
      operationId: create_menu_item_api_v1_menu_items_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menu/items/{item_id}:
    get:
      tags:
      - Menu & Catalog
      summary: Get menu item by ID
      description: Retrieve single menu item details.
      operationId: get_menu_item_api_v1_menu_items__item_id__get
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Menu & Catalog
      summary: Update menu item (admin only)
      description: Update an existing menu item (requires admin privileges).
      operationId: update_menu_item_api_v1_menu_items__item_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Menu & Catalog
      summary: Delete menu item (admin only)
      description: Delete a menu item from the catalog (requires admin privileges).
      operationId: delete_menu_item_api_v1_menu_items__item_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menu/items/{item_id}/availability:
    patch:
      tags:
      - Menu & Catalog
      summary: Toggle menu item availability in stop-list (admin only)
      description: Instantly toggle menu item availability in stop-list (requires
        admin privileges).
      operationId: toggle_item_availability_api_v1_menu_items__item_id__availability_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemAvailabilityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    CategoryCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 2
          title: Name
        slug:
          type: string
          maxLength: 100
          minLength: 2
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        display_order:
          type: integer
          title: Display Order
          default: 0
      type: object
      required:
      - name
      - slug
      title: CategoryCreate
    CategoryResponse:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 2
          title: Name
        slug:
          type: string
          maxLength: 100
          minLength: 2
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        display_order:
          type: integer
          title: Display Order
          default: 0
        id:
          type: integer
          title: Id
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - name
      - slug
      - id
      - is_active
      title: CategoryResponse
    CategoryUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 2
          - type: 'null'
          title: Name
        slug:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 2
          - type: 'null'
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      type: object
      title: CategoryUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ItemAvailabilityUpdate:
      properties:
        is_available:
          type: boolean
          title: Is Available
          description: Availability status in stop-list
      type: object
      required:
      - is_available
      title: ItemAvailabilityUpdate
...
    MenuItemCreate:
      properties:
        name:
          type: string
          maxLength: 150
          minLength: 2
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        price:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,8}|(?=[\d.]{1,11}0*$)\d{0,8}\.\d{0,2}0*$)
          title: Price
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        is_available:
          type: boolean
          title: Is Available
          default: true
        category_id:
          type: integer
          title: Category Id
      type: object
      required:
      - name
      - price
      - category_id
      title: MenuItemCreate
    MenuItemResponse:
      properties:
        name:
          type: string
          maxLength: 150
          minLength: 2
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,8}|(?=[\d.]{1,11}0*$)\d{0,8}\.\d{0,2}0*$)
          title: Price
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        is_available:
          type: boolean
          title: Is Available
          default: true
        id:
          type: integer
          title: Id
        category_id:
          type: integer
          title: Category Id
      type: object
      required:
      - name
      - price
      - id
      - category_id
      title: MenuItemResponse
    MenuItemUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 2
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        price:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,8}|(?=[\d.]{1,11}0*$)\d{0,8}\.\d{0,2}0*$)
          - type: 'null'
          title: Price
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        is_available:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Available
        category_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Category Id
      type: object
      title: MenuItemUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer